+++ src/interface/IArea.ts
@@ -1,0 +1,8 @@
+import { SplitAreaDirective } from "../components/splitArea.directive";
+
+export interface IArea {
+    comp: SplitAreaDirective;
+    size: number;
+    pxToSubtract: number;
+    order: number;
+}